+++ CHANGELOG.md
@@ -34,6 +34,7 @@ n/a
 * Removed `PasswordVaultBaseFragment` for newer `PasswordVaultFragment<V extends ViewModel>` which handles view models.
 * Remove deprecated recycler view adapter framework.
 * `MainActivity` displays a warning at the top of the page if a newer version of the app is available for download. This is done to immediately inform the user about a newer version.
+* Fixed bug where entries were sometimes not displayed after being loaded.
 
 <br/>